extract NodeActivator
