+++ cocos2d/core/component-scheduler.js
@@ -24,32 +24,19 @@
  ****************************************************************************/
 
 require('./platform/CCClass');
-var CCObject = require('./platform/CCObject');
+var Flags = require('./platform/CCObject').Flags;
 var JsArray = require('./platform/js').array;
 
-var Flags = CCObject.Flags;
 var IsStartCalled = Flags.IsStartCalled;
 var IsOnEnableCalled = Flags.IsOnEnableCalled;
 var IsEditorOnEnableCalled = Flags.IsEditorOnEnableCalled;
-var IsOnLoadStarted = Flags.IsOnLoadStarted;
-var IsOnLoadCalled = Flags.IsOnLoadCalled;
-var Deactivating = Flags.Deactivating;
 
-var callerFunctor;
-if (CC_EDITOR) {
-    callerFunctor = require('./utils/misc').tryCatchFunctor_EDITOR;
-}
-var callPreloadInTryCatch = CC_EDITOR && callerFunctor('__preload');
-var callOnLoadInTryCatch = CC_EDITOR && callerFunctor('onLoad');
+var callerFunctor = CC_EDITOR && require('./utils/misc').tryCatchFunctor_EDITOR;
 var callOnEnableInTryCatch = CC_EDITOR && callerFunctor('onEnable');
 var callStartInTryCatch = CC_EDITOR && callerFunctor('start', null, null, 'target._objFlags |= ' + IsStartCalled);
 var callUpdateInTryCatch = CC_EDITOR && callerFunctor('update', 'dt', 'dt');
 var callLateUpdateInTryCatch = CC_EDITOR && callerFunctor('lateUpdate', 'dt', 'dt');
 var callOnDisableInTryCatch = CC_EDITOR && callerFunctor('onDisable');
-var callOnDestroyInTryCatch = CC_EDITOR && callerFunctor('onDestroy');
-var callResetInTryCatch = CC_EDITOR && callerFunctor('resetInEditor');
-var callOnFocusInTryCatch = CC_EDITOR && callerFunctor('onFocusInEditor');
-var callOnLostFocusInTryCatch = CC_EDITOR && callerFunctor('onLostFocusInEditor');
 
 function sortedIndex (array, comp) {
     var order = comp.constructor._executionOrder;
@@ -82,6 +69,19 @@ function sortedIndex (array, comp) {
     return ~l;
 }
 
+function stableRemoveInactive (iterator) {
+    var array = iterator.array;
+    for (var i = 0; i < array.length;) {
+        var comp = array[i];
+        if (comp._enabled && comp.node._activeInHierarchy) {
+            ++i;
+        }
+        else {
+            iterator.removeAt(i);
+        }
+    }
+}
+
 // This class contains some queues used to invoke life-cycle methods by script execution order
 var LifeCycleInvoker = cc.Class({
     __ctor__ (invokeFunc) {
@@ -95,27 +95,16 @@ var LifeCycleInvoker = cc.Class({
 
         this._invoke = invokeFunc;
     },
+    statics: {
+        stableRemoveInactive
+    },
     add: null,
     remove: null,
     invoke: null,
-    sort (iterator) {
-        iterator.array.sort(function (a, b) {
-            return a.constructor._executionOrder - b.constructor._executionOrder;
-        });
-    },
 });
 
-function stableRemoveInactive (iterator) {
-    var array = iterator.array;
-    for (var i = 0; i < array.length;) {
-        var comp = array[i];
-        if (comp._enabled && comp.node._activeInHierarchy) {
-            ++i;
-        }
-        else {
-            iterator.removeAt(i);
-        }
-    }
+function compareOrder (a, b) {
+    return a.constructor._executionOrder - b.constructor._executionOrder;
 }
 
 // for onLoad: sort once all components registered, invoke once
@@ -137,7 +126,7 @@ var OneOffInvoker = cc.Class({
     invoke () {
         var compsNeg = this._neg;
         if (compsNeg.array.length > 0) {
-            this.sort(compsNeg);
+            compsNeg.array.sort(compareOrder);
             this._invoke(compsNeg);
             compsNeg.array.length = 0;
         }
@@ -147,7 +136,7 @@ var OneOffInvoker = cc.Class({
 
         var compsPos = this._pos;
         if (compsPos.array.length > 0) {
-            this.sort(compsPos);
+            compsPos.array.sort(compareOrder);
             this._invoke(compsPos);
             compsPos.array.length = 0;
         }
@@ -199,24 +188,6 @@ var ReusableInvoker = cc.Class({
     },
 });
 
-// for __preload: use internally, no sort
-var UnsortedInvoker = cc.Class({
-    extends: LifeCycleInvoker,
-    add (comp) {
-        this._zero.array.push(comp);
-    },
-    remove (comp) {
-        this._zero.fastRemove(comp);
-    },
-    cancelInactive () {
-        stableRemoveInactive(this._zero);
-    },
-    invoke () {
-        this._invoke(this._zero);
-        this._zero.array.length = 0;
-    },
-});
-
 function enableInEditor (comp) {
     if (!(comp._objFlags & IsEditorOnEnableCalled)) {
         cc.engine.emit('component-enabled', comp.uuid);
@@ -224,7 +195,7 @@ function enableInEditor (comp) {
     }
 }
 
-function getInvokeImpl (code, useDt) {
+function createInvokeImpl (code, useDt) {
     if (CC_EDITOR) {
         if (useDt) {
             return function (iterator) {
@@ -255,8 +226,8 @@ function getInvokeImpl (code, useDt) {
         // }
         var body = 'var a=it.array;' +
                    'for(it.i=0;it.i<a.length;++it.i){' +
-                       'var c=a[it.i];' +
-                        code +
+                   'var c=a[it.i];' +
+                   code +
                    '}';
         if (useDt) {
             return new Function('it', 'dt', body);
@@ -267,110 +238,96 @@ function getInvokeImpl (code, useDt) {
     }
 }
 
-function createActivateTask () {
-    var preload = new UnsortedInvoker(getInvokeImpl(
-        CC_EDITOR ? callPreloadInTryCatch : 'c.__preload();'));
 
-    var onLoad = new OneOffInvoker(CC_EDITOR ? function (iterator) {
-        var array = iterator.array;
-        for (iterator.i = 0; iterator.i < array.length; ++iterator.i) {
-            let comp = array[iterator.i];
-            callOnLoadInTryCatch(comp);
-            if (!comp.node._activeInHierarchy) {
-                // deactivated during onLoad
-                break;
-            }
-        }
-    } : getInvokeImpl('c.onLoad();if(!c.node._activeInHierarchy)break;'));
-
-    var onEnable = new OneOffInvoker(CC_EDITOR ? function (iterator) {
-        var scheduler = cc.director.getScheduler();
-        var compScheduler = cc.director._compScheduler;
-        var array = iterator.array;
-        for (iterator.i = 0; iterator.i < array.length; ++iterator.i) {
-            let comp = array[iterator.i];
-            if (comp._enabled) {
-                callOnEnableInTryCatch(comp);
-                var deactivatedDuringOnEnable = !comp.node._activeInHierarchy;
-                if (!deactivatedDuringOnEnable) {
-                    scheduler.resumeTarget(comp);
-                    compScheduler.schedule(comp);
-                    comp._objFlags |= IsOnEnableCalled;
-                }
-            }
-        }
-    } : function (iterator) {
-        var scheduler = cc.director.getScheduler();
-        var compScheduler = cc.director._compScheduler;
-        var array = iterator.array;
-        for (iterator.i = 0; iterator.i < array.length; ++iterator.i) {
-            let comp = array[iterator.i];
-            if (comp._enabled) {
-                comp.onEnable();
-                var deactivatedDuringOnEnable = !comp.node._activeInHierarchy;
-                if (!deactivatedDuringOnEnable) {
-                    scheduler.resumeTarget(comp);
-                    compScheduler.schedule(comp);
-                    comp._objFlags |= IsOnEnableCalled;
-                }
-            }
-        }
-    });
-
-    return { preload, onLoad, onEnable };
-}
-
-var MAX_POOL_SIZE = 4;
-var activateTasksPool = [];
-// get invoker temporary
-function getActivateTask () {
-    return activateTasksPool.pop() || createActivateTask();
-}
-// release invoker temporary
-function putActivateTask (task) {
-    if (activateTasksPool.length < MAX_POOL_SIZE) {
-        activateTasksPool.push(task);
-    }
-}
-
+/**
+ * The Manager for Component's life-cycle methods.
+ */
 function ctor () {
     // invokers
-    this.startInvoker = new OneOffInvoker(getInvokeImpl(
+    this.startInvoker = new OneOffInvoker(createInvokeImpl(
         CC_EDITOR ? callStartInTryCatch : 'c.start();c._objFlags|=' + IsStartCalled));
-    this.updateInvoker = new ReusableInvoker(getInvokeImpl(
+    this.updateInvoker = new ReusableInvoker(createInvokeImpl(
         CC_EDITOR ? callUpdateInTryCatch : 'c.update(dt)', true));
-    this.lateUpdateInvoker = new ReusableInvoker(getInvokeImpl(
+    this.lateUpdateInvoker = new ReusableInvoker(createInvokeImpl(
         CC_EDITOR ? callLateUpdateInTryCatch : 'c.lateUpdate(dt)', true));
 
     // components deferred to next frame
     this.scheduleInNextFrame = [];
 
-    // a stack of node's activating tasks
-    this._activatingStack = [];
-
     // during a loop
     this._updating = false;
 }
-
-function _componentCorrupted (node, comp, index) {
-    if (CC_DEV) {
-        cc.errorID(3817, node.name, index);
-        console.log('Corrupted component value:', comp);
-    }
-    if (comp) {
-        node._removeComponent(comp);
-    }
-    else {
-        JS.array.removeAt(node._components, index);
-    }
-}
-
-/**
- * The Manager for Component's life-cycle methods.
- */
 var ComponentScheduler = cc.Class({
     ctor: ctor,
     unscheduleAll: ctor,
+
+    statics: {
+        LifeCycleInvoker,
+        OneOffInvoker,
+        createInvokeImpl,
+        invokeOnEnable: CC_EDITOR ? function (iterator) {
+            var compScheduler = cc.director._compScheduler;
+            var array = iterator.array;
+            for (iterator.i = 0; iterator.i < array.length; ++iterator.i) {
+                let comp = array[iterator.i];
+                if (comp._enabled) {
+                    callOnEnableInTryCatch(comp);
+                    var deactivatedDuringOnEnable = !comp.node._activeInHierarchy;
+                    if (!deactivatedDuringOnEnable) {
+                        compScheduler._onEnabled(comp);
+                    }
+                }
+            }
+        } : function (iterator) {
+            var compScheduler = cc.director._compScheduler;
+            var array = iterator.array;
+            for (iterator.i = 0; iterator.i < array.length; ++iterator.i) {
+                let comp = array[iterator.i];
+                if (comp._enabled) {
+                    comp.onEnable();
+                    var deactivatedDuringOnEnable = !comp.node._activeInHierarchy;
+                    if (!deactivatedDuringOnEnable) {
+                        compScheduler._onEnabled(comp);
+                    }
+                }
+            }
+        }
+    },
+
+    _onEnabled (comp) {
+        cc.director.getScheduler().resumeTarget(comp);
+        comp._objFlags |= IsOnEnableCalled;
+
+        // schedule
+        if (this._updating) {
+            this.scheduleInNextFrame.push(comp);
+            return;
+        }
+        this._scheduleImmediate(comp);
+    },
+
+    _onDisabled (comp) {
+        cc.director.getScheduler().pauseTarget(comp);
+        comp._objFlags &= ~IsOnEnableCalled;
+
+        // cancel schedule task
+        var index = this.scheduleInNextFrame.indexOf(comp);
+        if (index >= 0) {
+            JsArray.fastRemoveAt(this.scheduleInNextFrame, index);
+            return;
+        }
+
+        // unschedule
+        if (comp.start && !(comp._objFlags & IsStartCalled)) {
+            this.startInvoker.remove(comp);
+        }
+        if (comp.update) {
+            this.updateInvoker.remove(comp);
+        }
+        if (comp.lateUpdate) {
+            this.lateUpdateInvoker.remove(comp);
+        }
+    },
 
     enableComp: CC_EDITOR ? function (comp, invoker) {
         if (cc.engine.isPlaying || comp.constructor._executeInEditMode) {
@@ -390,10 +347,7 @@ var ComponentScheduler = cc.Class({
                         }
                     }
                 }
-                cc.director.getScheduler().resumeTarget(comp);
-                this.schedule(comp);
-                
-                comp._objFlags |= IsOnEnableCalled;
+                this._onEnabled(comp);
             }
         }
         enableInEditor(comp);
@@ -413,10 +367,7 @@ var ComponentScheduler = cc.Class({
                     }
                 }
             }
-            cc.director.getScheduler().resumeTarget(comp);
-            this.schedule(comp);
-
-            comp._objFlags |= IsOnEnableCalled;
+            this._onEnabled(comp);
         }
     },
 
@@ -426,11 +377,7 @@ var ComponentScheduler = cc.Class({
                 if (comp.onDisable) {
                     callOnDisableInTryCatch(comp);
                 }
-
-                cc.director.getScheduler().pauseTarget(comp);
-                this.unschedule(comp);
-
-                comp._objFlags &= ~IsOnEnableCalled;
+                this._onDisabled(comp);
             }
         }
         if (comp._objFlags & IsEditorOnEnableCalled) {
@@ -442,247 +389,7 @@ var ComponentScheduler = cc.Class({
             if (comp.onDisable) {
                 comp.onDisable();
             }
-
-            cc.director.getScheduler().pauseTarget(comp);
-            this.unschedule(comp);
-
-            comp._objFlags &= ~IsOnEnableCalled;
-        }
-    },
-
-    _activateNodeRecursively (node, preloadInvoker, onLoadInvoker, onEnableInvoker) {
-        if (node._objFlags & Deactivating) {
-            // en:
-            // Forbid reactive the same node during its deactivating procedure
-            // to avoid endless loop and simplify the implementation.
-            // zh:
-            // 对相同节点而言，无法撤销反激活，防止反激活 - 激活 - 反激活的死循环发生。
-            // 这样设计简化了一些引擎的实现，而且对调用者来说能保证反激活操作都能成功。
-            cc.errorID(3816, node.name);
-            return;
-        }
-
-        node._activeInHierarchy = true;
-
-        // component maybe added during onEnable, and the onEnable of new component is already called
-        // so we should record the origin length
-        var originCount = node._components.length;
-        // activate components
-        for (let i = 0; i < originCount; ++i) {
-            let component = node._components[i];
-            if (component instanceof cc.Component) {
-                this.activateComp(component, preloadInvoker, onLoadInvoker, onEnableInvoker);
-            }
-            else {
-                _componentCorrupted(node, component, i);
-                --i;
-                --originCount;
-            }
-        }
-        // activate children recursively
-        for (let i = 0, len = node._children.length; i < len; ++i) {
-            let child = node._children[i];
-            if (child._active) {
-                this._activateNodeRecursively(child, preloadInvoker, onLoadInvoker, onEnableInvoker);
-            }
-        }
-
-        node._onPostActivated(true);
-    },
-
-    _deactivateNodeRecursively (node) {
-        if (CC_TEST || CC_DEV) {
-            cc.assert(!(node._objFlags & Deactivating), 'node should not deactivating');
-            // ensures _activeInHierarchy is always changing when Deactivating flagged
-            cc.assert(node._activeInHierarchy, 'node should not deactivated');
-        }
-        node._objFlags |= Deactivating;
-        node._activeInHierarchy = false;
-
-        // component maybe added during onEnable, and the onEnable of new component is already called
-        // so we should record the origin length
-        var originCount = node._components.length;
-        for (let c = 0; c < originCount; ++c) {
-            let component = node._components[c];
-            if (component._enabled) {
-                this.disableComp(component);
-
-                if (node._activeInHierarchy) {
-                    // reactivated from root
-                    node._objFlags &= ~Deactivating;
-                    return;
-                }
-            }
-        }
-        for (let i = 0, len = node._children.length; i < len; ++i) {
-            let child = node._children[i];
-            if (child._activeInHierarchy) {
-                this._deactivateNodeRecursively(child);
-
-                if (node._activeInHierarchy) {
-                    // reactivated from root
-                    node._objFlags &= ~Deactivating;
-                    return;
-                }
-            }
-        }
-
-        node._onPostActivated(false);
-        node._objFlags &= ~Deactivating;
-    },
-
-    activateNode (node, active) {
-        if (active) {
-            var task = getActivateTask(node);
-            this._activatingStack.push(task);
-
-            this._activateNodeRecursively(node, task.preload, task.onLoad, task.onEnable);
-            task.preload.invoke();
-            task.onLoad.invoke();
-            task.onEnable.invoke();
-
-            this._activatingStack.pop();
-            putActivateTask(task);
-        }
-        else {
-            this._deactivateNodeRecursively(node);
-
-            // remove children of this node from previous activating tasks to debounce
-            // (this is an inefficient operation but it ensures general case could be implemented in a efficient way)
-            var stack = this._activatingStack;
-            for (var i = 0; i < stack.length; i++) {
-                var lastTask = stack[i];
-                lastTask.preload.cancelInactive(node);
-                lastTask.onLoad.cancelInactive(node);
-                lastTask.onEnable.cancelInactive(node);
-            }
-        }
-        node.emit('active-in-hierarchy-changed', node);
-    },
-
-    activateComp: CC_EDITOR ? function (comp, preloadInvoker, onLoadInvoker, onEnableInvoker) {
-        if (!(comp._objFlags & IsOnLoadStarted) &&
-            (cc.engine._isPlaying || comp.constructor._executeInEditMode)) {
-            comp._objFlags |= IsOnLoadStarted;
-
-            if (typeof comp.__preload === 'function') {
-                if (preloadInvoker) {
-                    preloadInvoker.add(comp);
-                }
-                else {
-                    callPreloadInTryCatch(comp);
-                }
-            }
-            if (comp.onLoad) {
-                if (onLoadInvoker) {
-                    onLoadInvoker.add(comp);
-                }
-                else {
-                    callOnLoadInTryCatch(comp);
-                }
-            }
-
-            comp._objFlags |= IsOnLoadCalled;
-
-            if (comp.onLoad && !cc.engine._isPlaying) {
-                var focused = Editor.Selection.curActivate('node') === comp.node.uuid;
-                if (focused && comp.onFocusInEditor) {
-                    callOnFocusInTryCatch(comp);
-                }
-                else if (comp.onLostFocusInEditor) {
-                    callOnLostFocusInTryCatch(comp);
-                }
-            }
-            if ( !CC_TEST ) {
-                _Scene.AssetsWatcher.start(comp);
-            }
-        }
-        if (comp._enabled) {
-            var deactivatedOnLoading = !comp.node._activeInHierarchy;
-            if (deactivatedOnLoading) {
-                return;
-            }
-            this.enableComp(comp, onEnableInvoker);
-        }
-    } : function (comp, preloadInvoker, onLoadInvoker, onEnableInvoker) {
-        if (!(comp._objFlags & IsOnLoadStarted)) {
-            comp._objFlags |= IsOnLoadStarted;
-
-            if (typeof comp.__preload === 'function') {
-                if (preloadInvoker) {
-                    preloadInvoker.add(comp);
-                }
-                else {
-                    comp.__preload();
-                }
-            }
-            if (comp.onLoad) {
-                if (onLoadInvoker) {
-                    onLoadInvoker.add(comp);
-                }
-                else {
-                    comp.onLoad();
-                }
-            }
-
-            comp._objFlags |= IsOnLoadCalled;
-        }
-        if (comp._enabled) {
-            var deactivatedOnLoading = !comp.node._activeInHierarchy;
-            if (deactivatedOnLoading) {
-                return;
-            }
-            this.enableComp(comp, onEnableInvoker);
-        }
-    },
-
-    destroyComp: CC_EDITOR ? function (comp) {
-        // ensure onDisable called
-        this.disableComp(comp);
-
-        if (comp.onDestroy && (comp._objFlags & IsOnLoadCalled)) {
-            if (cc.engine._isPlaying || comp.constructor._executeInEditMode) {
-                callOnDestroyInTryCatch(comp);
-            }
-        }
-    } : function (comp) {
-        // ensure onDisable called
-        this.disableComp(comp);
-
-        if (comp.onDestroy && (comp._objFlags & IsOnLoadCalled)) {
-            comp.onDestroy();
-        }
-    },
-
-    resetComp: CC_EDITOR && function (comp) {
-        if (typeof comp.resetInEditor === 'function') {
-            callResetInTryCatch(comp);
-        }
-    },
-
-    schedule (comp) {
-        if (this._updating) {
-            this.scheduleInNextFrame.push(comp);
-            return;
-        }
-        this._scheduleImmediate(comp);
-    },
-    unschedule (comp) {
-        // cancel schedule task
-        var index = this.scheduleInNextFrame.indexOf(comp);
-        if (index >= 0) {
-            JsArray.fastRemoveAt(this.scheduleInNextFrame, index);
-            return;
-        }
-        // unschedule
-        if (comp.start && !(comp._objFlags & IsStartCalled)) {
-            this.startInvoker.remove(comp);
-        }
-        if (comp.update) {
-            this.updateInvoker.remove(comp);
-        }
-        if (comp.lateUpdate) {
-            this.lateUpdateInvoker.remove(comp);
+            this._onDisabled(comp);
         }
     },
